Split PdfCharacterCountToAbsolute in two classes
